Updated to use json and working lighting render task

+++ src/lib/ingestor/Task/Shotgun/SGPublish.py
@@ -1,8 +1,9 @@
 import os
-import subprocess
 import tempfile
 from ..Task import Task
 from ...Template import Template
+from ...Crawler.Fs.Path import Path
+from ...Crawler.Fs.Image import Image
 
 class SGPublish(Task):
     """
@@ -25,19 +26,31 @@ class SGPublish(Task):
         from ushotgun import Session
         sg = Session.get()
 
+        # Source crawler is a json crawler that points to published data
         sourceCrawler = self.pathCrawlers()[0]
 
-        self.__publishData["path"] = {"local_path": sourceCrawler.var('filePath')}
-        self.__publishData["code"] = sourceCrawler.var('name')
+        filePath = sourceCrawler.var('filePath')
+        # if sourceCrawler.isSequence():
+        #     filePath = os.path.join(
+        #         os.path.dirname(filePath),
+        #         '{0}.%0{1}d.{2}'.format(
+        #             sourceCrawler.var('name'),
+        #             sourceCrawler.var('padding'),
+        #             sourceCrawler.var('ext')
+        #             )
+        #         )
+
+        self.__publishData["path"] = {"local_path": filePath}
 
         self.__publishData["description"] = self.option('comment')
         self.__publishData["version_number"] = self.pathCrawlers()[0].var('version')
 
-        self.__publishData["task"] = None
         if "_sgTask" in sourceCrawler.varNames():
             self.__publishData["task"] = sourceCrawler.var("_sgTask")
 
-        self.__publishData["name"] = Template(self.option('publishName')).valueFromCrawler(sourceCrawler)
+        publishName = Template(self.option('publishName')).valueFromCrawler(sourceCrawler)
+        self.__publishData["name"] = publishName
+        self.__publishData["code"] = publishName
 
         self.__linkData(sg)
         self.__sgFileType(sg)
@@ -46,8 +59,6 @@ class SGPublish(Task):
         sgPublishFile = sg.create("PublishedFile", self.__publishData)
         self.__makeThumbnail(sgPublishFile, sg)
         self.__makeDaily(sgPublishFile, sg)
-
-        # Add dependencies
 
         # this task does not return any crawlers as result
         return []
@@ -97,12 +108,16 @@ class SGPublish(Task):
 
     def __makeThumbnail(self, sgPublishFile, sg):
 
+        createThumbnail = False
+        sourceCrawler = self.pathCrawlers()[0]
         if "thumbnailFile" in self.optionNames():
-            thumbnailFilePath = self.option('thumbnailFile')
+            thumbnailFilePath = Template(self.option('thumbnailFile')).valueFromCrawler(sourceCrawler)
         else:
-            targetCrawler = self.pathCrawlers()[int(len(self.pathCrawlers()) / 2)]
-            if targetCrawler.var("category") != "image":
+            createThumbnail = True
+            imageCrawlers = sourceCrawler.glob(filterTypes=[Image])
+            if not imageCrawlers:
                 return
+            targetCrawler = imageCrawlers[int(len(self.pathCrawlers()) / 2)]
 
             tempFile = tempfile.NamedTemporaryFile(
                 delete=False,
@@ -113,26 +128,37 @@ class SGPublish(Task):
             # Remove file so ffmpeg doesn't ask to overwrite it
             os.unlink(thumbnailFilePath)
 
+            # thumbnailTask = Task.create('imageThumbnail')
+            # thumbnailTask.add(targetCrawler, thumbnailFilePath)
+            # thumbnailTask.output()
             cmd = 'ffmpeg -v quiet -loglevel error -i {} -vf scale=640:-1 {}'.format(
                     targetCrawler.var("filePath"),
                     thumbnailFilePath
-                )
+                    )
             os.system(cmd)
 
         if os.path.exists(thumbnailFilePath):
             sg.upload_thumbnail("PublishedFile", sgPublishFile["id"], thumbnailFilePath)
 
-        # removing the temporary file
-        os.unlink(thumbnailFilePath)
+        if createThumbnail:
+            # removing the temporary file
+            os.unlink(thumbnailFilePath)
 
     def __makeDaily(self, sgPublishFile, sg):
         """
         Create a version in Shotgun for this path and linked to this publish.
         """
         sourceCrawler = self.pathCrawlers()[0]
-        movieFilePath = Template(self.option('movieFile')).valueFromCrawler(sourceCrawler)
-        if not movieFilePath or not os.path.exists(movieFilePath):
-            return
+        if 'movieFile' not in self.optionNames():
+            # No movie provided, glob for a mov
+            movCrawlers = sourceCrawler.glob(fileTypes=["mov"])
+            if not movCrawlers:
+                return
+            movieFilePath = movCrawlers[0].var("filePath")
+        else:
+            movieFilePath = Template(self.option('movieFile')).valueFromCrawler(sourceCrawler)
+            if not movieFilePath or not os.path.exists(movieFilePath):
+                raise Exception("Movie provided for daily creation does not exist: {}".format(movieFilePath))
 
         # create a name for the version based on the file name
         # grab the file name, strip off extension
@@ -142,24 +168,35 @@ class SGPublish(Task):
         # and capitalize
         name = name.capitalize()
 
-        firstFrame = self.pathCrawlers()[0].var('frame')
-        lastFrame = self.pathCrawlers()[-1].var('frame')
+        firstFrame = None
+        lastFrame = None
+        movCrawler = Path.createFromPath(movieFilePath)
+        if firstFrame in movCrawler.varNames():
+            firstFrame = movCrawler.var('firstFrame')
+            lastFrame = movCrawler.var('lastFrame')
+        # else:
+        #     imageCrawlers = sourceCrawler.glob(filterTypes=["Image"])
+        #     if imageCrawlers and imageCrawlers[0].isSequence():
+        #         firstFrame = self.pathCrawlers()[0].var('frame')
+        #         lastFrame = self.pathCrawlers()[-1].var('frame')
 
         # Create the version in Shotgun
         data = {
             "code": name,
             "sg_status_list": "rev",
             "entity": self.__publishData['entity'],
-            "sg_first_frame": firstFrame,
-            "sg_last_frame": lastFrame,
-            "frame_count": (lastFrame - firstFrame + 1),
-            "frame_range": "%s-%s" % (firstFrame, lastFrame),
             "created_by": self.__publishData['created_by'],
             "user": self.__publishData['created_by'],
             "description": self.__publishData['description'],
             "sg_path_to_frames": self.__publishData["path"]["local_path"],
             "project": self.__publishData['project']
         }
+
+        if firstFrame is not None and lastFrame is not None:
+            data["sg_first_frame"] = firstFrame
+            data["sg_last_frame"] = lastFrame
+            data["frame_count"] = (lastFrame - firstFrame + 1)
+            data["frame_range"] = "%s-%s" % (firstFrame, lastFrame)
 
         data["published_files"] = [sgPublishFile]
         data["sg_path_to_movie"] = movieFilePath
